Fix crash

+++ Frisbee/Framework/ID.swift
@@ -8,6 +8,6 @@
 
 extension Int {
     static var randomID: Int {
-        return Int(arc4random_uniform(UInt32(Int.max)))
+        return Int(arc4random_uniform(UInt32.max))
     }
 }
